Lesson 5 updated
so far as I am in lesson five of the youtube tutorial I have changed things to work so that everything in it will work the way it is currently written and meant to be learned. App.js will stay as it is for now, but a text file with the copied code will be in there too.